fix: address react doctor feedback
