+++ apps/web/src/routes/_authenticated/hair-sales/$hairSaleId.tsx
@@ -6,13 +6,13 @@ import { trpc } from "@/utils/trpc"
 import { HairSaleDetailPage } from "./-components/hair-sale-id-page"
 
 export const Route = createFileRoute("/_authenticated/hair-sales/$hairSaleId")({
-  component: routeComponent,
+  component: RouteComponent,
   loader: async ({ context, params }) => {
     await context.queryClient.ensureQueryData(trpc.hairAssigned.get.queryOptions({ id: params.hairSaleId }))
   },
 })
 
-function routeComponent() {
+function RouteComponent() {
   const { hairSaleId } = Route.useParams()
   const sale = useQuery(trpc.hairAssigned.get.queryOptions({ id: hairSaleId })).data
