+++ apps/web/src/routes/_authenticated/legal-entities/$legalEntityId/salons.tsx
@@ -6,10 +6,10 @@ import { trpc } from "@/utils/trpc"
 import { SalonsTab } from "./-components/salons-page"
 
 export const Route = createFileRoute("/_authenticated/legal-entities/$legalEntityId/salons")({
-  component: routeComponent,
+  component: RouteComponent,
 })
 
-function routeComponent() {
+function RouteComponent() {
   const salonsData = useQuery(trpc.salons.list.queryOptions({ pageSize: 100 })).data
 
   return <SalonsTab salons={salonsData?.items ?? []} />
