+++ apps/web/src/routes/_authenticated/legal-entities/index.tsx
@@ -6,10 +6,10 @@ import { trpc } from "@/utils/trpc"
 import { LegalEntitiesIndex } from "./-components/index-page"
 
 export const Route = createFileRoute("/_authenticated/legal-entities/")({
-  component: routeComponent,
+  component: RouteComponent,
 })
 
-function routeComponent() {
+function RouteComponent() {
   const legalEntitiesQuery = useQuery(trpc.legalEntities.list.queryOptions({ pageSize: 100 }))
   const unassignedAttachments = useQuery(
     trpc.bankStatementAttachments.list.queryOptions({ assignmentStatus: "unassigned" }),
